+++ CHANGELOG.md
@@ -4,6 +4,9 @@ A breaking change will get clearly marked in this log.
 
 ## Unreleased
 
+### Fix
+- Utils.readTransactionTx now checks timebounds with a 5-minute grace period to account for clock drift.
+
 ## [v8.2.3](https://github.com/stellar/js-stellar-sdk/compare/v8.2.2...v8.2.3)
 
 ### Fix